iot_platform3/edge-rpi-v2/plc-collector/requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0.2
python-dotenv==1.0.1
python-json-logger==2.0.7
